refactor: Location 객체 합치기
#43
