+++ app/src/main/java/com/boosters/promise/data/promise/Promise.kt
@@ -1,8 +1,9 @@
 package com.boosters.promise.data.promise
 
-import com.boosters.promise.ui.promise.model.PromiseUiState
+import com.boosters.promise.data.model.Location
 import com.boosters.promise.data.user.User
 import com.boosters.promise.data.user.toUserUiState
+import com.boosters.promise.ui.promise.model.PromiseUiState
 
 data class Promise(
     val promiseId: String,
