+++ app/src/main/java/com/boosters/promise/data/promise/PromiseBody.kt
@@ -1,5 +1,6 @@
 package com.boosters.promise.data.promise
 
+import com.boosters.promise.data.model.Location
 import com.boosters.promise.data.user.User
 
 data class PromiseBody(
