--- app/src/main/java/com/boosters/promise/data/user/Location.kt
@@ -1,6 +1,0 @@
-package com.boosters.promise.data.user
-
-data class Location(
-    val x: Int? = null,
-    val y: Int? = null
-)
